Refactor bra size measurement
